extract view attributes related functions to ViewAttributesAction.js
https://github.com/metasfresh/metasfresh-webui-frontend/issues/1357

@wiadev  pls review
